+++ wsd/LOOLWSD.cpp
@@ -261,7 +261,7 @@ inline void checkSessionLimitsAndWarnClients()
 #if !MOBILEAPP
     ssize_t docBrokerCount = DocBrokers.size() - ConvertToBroker::getInstanceCount();
     if (LOOLWSD::MaxDocuments < 10000 &&
-        (docBrokerCount > LOOLWSD::MaxDocuments || LOOLWSD::NumConnections >= LOOLWSD::MaxConnections))
+        (docBrokerCount > static_cast<ssize_t>(LOOLWSD::MaxDocuments) || LOOLWSD::NumConnections >= LOOLWSD::MaxConnections))
     {
         const std::string info = Poco::format(PAYLOAD_INFO_LIMIT_REACHED, LOOLWSD::MaxDocuments, LOOLWSD::MaxConnections);
         LOG_INF("Sending client 'limitreached' message: " << info);
